requirements.txt
selenium
beautifulsoup4
requests
pandas
flask
openai
pyyaml
